updated README.md
Included simple commands to help new users automate their install, upgrade, and upgrade-dist tasks in Kali. Would like to build this out to include extra tools and the option to select from options that save to a file locally and get called to run the current sequence from boot and update and run tasks silently and automatically in the background.